[core][clickhouse] move clickhouse-specific sql to clickhouse crate
