+++ tesseract-clickhouse/Cargo.toml
@@ -8,6 +8,7 @@ version = "0.1.0"
 clickhouse-rs = "0.1.5"
 failure = "0.1.2"
 futures = "0.1.25"
+itertools = "0.8.0"
 log = "0.4.3"
 
 [dependencies.tesseract-core]
