+++ tesseract-clickhouse/src/sql/growth.rs
